pyproject.toml
[tool.poetry]
name = "cor-lab-local"
version = "0.1.0"
description = ""
authors = ["ZhovtenkoA <zhowtenko.oleksiy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.1"
alembic = "^1.16.4"
redis = "^6.2.0"
passlib = "^1.7.4"
uvicorn = "^0.35.0"
sqlalchemy = "^2.0.41"
pydantic = "^2.11.7"
python-jose = "^3.5.0"
python-multipart = "^0.0.20"
fastapi-mail = "^1.5.0"
httpx = "^0.28.1"
pycryptodome = "^3.23.0"
black = "^25.1.0"
pydantic-settings = "^2.10.1"
qrcode = "^8.2"
pillow = "^11.3.0"
cryptography = "^45.0.5"
requests = "^2.32.4"
prometheus-client = "^0.22.1"
loguru = "^0.7.3"
pydicom = "^3.0.1"
asyncpg = "^0.30.0"
numpy = "^2.3.1"
scikit-image = "^0.25.2"
openslide-python = "^1.4.2"
openslide-bin = "^4.0.0.8"
pip = "^25.1.1"
prometheus-fastapi-instrumentator = "^7.1.0"
gdcm = "^1.1"
pylibjpeg = "^2.0.1"
pylibjpeg-libjpeg = "^2.3.0"
pylibjpeg-openjpeg = "^2.4.0"
gunicorn = "^23.0.0"
fastapi-limiter = "^0.1.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
